Part #11: Access Control (user can edit & delete his own posts only)

+++ app/Http/Controllers/PostsController.php
@@ -9,6 +9,18 @@ use DB;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', [
+          'except' => ['index', 'show']
+        ]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -83,6 +95,11 @@ class PostsController extends Controller
     {
       $post = Post::find($id);
 
+      // Check for correct user
+      if ( auth()->user()->id != $post->user_id ) {
+        return redirect('/posts')->with('error', 'Unauthorized post!');
+      }
+
       return view('posts.edit')->with('post', $post);
     }
 
@@ -102,6 +119,12 @@ class PostsController extends Controller
 
       // Create post
       $post = Post::find($id);
+
+      // Check for correct user
+      if ( auth()->user()->id != $post->user_id ) {
+        return redirect('/posts')->with('error', 'Unauthorized post!');
+      }
+
       $post->title = $request->input('title');
       $post->body = $request->input('body');
       $post->save();
@@ -118,6 +141,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
       $post = Post::find($id);
+
+      // Check for correct user
+      if ( auth()->user()->id != $post->user_id ) {
+        return redirect('/posts')->with('error', 'Unauthorized post!');
+      }
+
       $post->delete();
 
       return redirect('/posts')->with('success', 'Post removed!');
